requirements.txt
beautifulsoup4==4.12.2
fake-useragent==1.2.1
fastapi==0.103.1
lxml==4.9.3
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.23.2

fastapi-users==12.1.2
# устанавливать из терминала pip install fastapi-users[sqlalchemy]
aiosqlite==0.19.0
SQLAlchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0